Move domain related classes to separate assembly
