+++ MyChat.Domain/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/obj/
