+++ MyChat.Domain/Filters/KeywordFilter.cs
@@ -1,4 +1,4 @@
-namespace MindLink.MyChat.Filters
+namespace MindLink.MyChat.Domain.Filters
 {
     public class KeywordFilter : IMessageFilter
     {
