+++ MyChat.Domain/Transformers/PhoneNumberTransformer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
 
-namespace MindLink.MyChat.Transformers
+namespace MindLink.MyChat.Domain.Transformers
 {
     public class PhoneNumberTransformer : IMessageTransformer
     {
