+++ MyChat.Domain/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="libphonenumber-csharp" version="8.3.1.1" targetFramework="net461" />
+  <package id="murmurhash" version="1.0.0" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net461" />
+</packages>
